refactor: add setup checkout fixture - this fixture adds cookies and data relevant to checkout to the browser - moved `load_json` and `camel_to_snake` functions to common utils file
